third_party/blink/web_tests/platform/generic/svg/dynamic-updates/SVGFEColorMatrixElement-svgdom-values-prop-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'values' property of the SVGFEColorMatrixElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS SVGNumberListToString(matrixElement.values.baseVal) is "0.9 0.9 0.9 0 0 0.33 0.33 0.33 0 0 0.33 0.33 0.33 0 0 0.33 0.33 0.33 0 0 "
PASS Math.round(saturateElement.values.baseVal.getItem(0).value * 1000) is 500
PASS Math.round(hueRotateElement.values.baseVal.getItem(0).value * 1000) is 10000
PASS SVGNumberListToString(matrixElement.values.baseVal) is "0.33 0.33 0.33 0 0 0.33 0.33 0.33 0 0 0.33 0.33 0.33 0 0 0.33 0.33 0.33 0 0 "
PASS saturateElement.values.baseVal.getItem(0).value is 0.25
PASS hueRotateElement.values.baseVal.getItem(0).value is 90
PASS successfullyParsed is true

TEST COMPLETE
